Moving fix to CvPlayer::SetTurnActive
The intention here is to facilitate testing within the SetTurnActive function, which seems closer to the root of the issue (sending gDLL->sendTurnUnready() before SetTurnActive
does not fix the issue, sending it after fixes the issue).

Also think the issue is best described as the AI playing twice in one turn, rather then the player losing one turn, because the turn count does not increase when the AI plays twice.